Re-arranged the packages + tests so that the methods that were shared are all foudn in one python file (less spaghettiing)--> sharedmethods.py
